Bug 1280: Ensure list parent if written item is list item
One of properties of Java Binding is that it is imposible
to represent list as a whole and thus it was possible
to create put operation, which was correct by Binding Spec
but failed.

This fixes behaviour, if user is writing list item (MapEntryNode),
and whole list (MapNode) does not exists, list is automatically
added.

Change-Id: I41e0e30d349720f50697a6f555bab5850ec7d30c
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>