Bug 4145: Issue augmenting choice inside a list
The problem occurs when target node of uses-augment statement
is added by uses from another grouping and the uses statement
is not directly inside the grouping, but it is nested below
other elements (e.g. list, container etc.) in the grouping.
In this case the original of the target node is not found and
the IllegalStateException is thrown.

Change-Id: I1b0329d5da754251040894c0ccfe920bf8190aa4
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>