Handle null BindingRuntimeContext in BindingToNormalizedNodeCodec
In the getModuleBlocking method, if timing is right, runtimeContext may
be null when it is accessed to get the SchemaContext. It is set in the
onGlobalContextUpdated method which may not have been called yet prior
to getModuleBlocking. There's already code to wait for schema via the
futureSchema so it shold check for null runtimeContext as well. Also
runtimeContext shoild be volatile as it's accessed by multiple threads.

Change-Id: I335b0c359d58f3c58da34c7255b75ac793e49bd6
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>