BUG-5414 introduce EOS inJeopardy flag
The inJeopardy flag is used to indicate that the leader has lost quorum,
e.g. if cannot reach majority of followers or the follower has lost connection
to the leader (and has initiated new elections).

While EOS is in jeopardy, any reported entity state may not reflect cluster-wide
consensus, but rather represents the latest intended state as seen by this node.

Depend on:
https://git.opendaylight.org/gerrit/#/c/36752/7

Change-Id: I8f17fb97dd6ec54222b783bb7b4b827ce9b4a0ad
Signed-off-by: Vaclav Demcak <vdemcak@cisco.com>