Update patch set 2
Patch Set 2:

ClassLoaderUtils was searching for a home for a loong time, at some point there actually were two versions with slightly different functionality. This one was not removed originally to keep API compatibility (when binding was still in yangtools) and we forgot to delete it when migrating code. Thanks for the patch.

Patch-set: 2