Bug 2872: Generated Java Enumerations should contain mapping to the string counter part
Changes the Enumeration API to expose the "getRawName()" accessor, which exposes the
name which was extracted directly from the YANG source.  This is useful for reverse
engineering java binding code to yang source code.

A corresponding change in yangtools, which must be merged first, exposes the getRawName()
method from the EnumPair API.

Change-Id: If248039a2b10958ff563ee0799d7b13284570b3c
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>