Bug 2872: Generated Java Enumerations should contain mapping to the string counter part
Changes the Enumeration API to expose the "getMappedName()" accessor, which exposes
the name that is used for the Java mapped enum name.  The EnumTemplate.xtend template
was modified in the following way:

1) A field was added to track the "raw" enum name, the one pared from the YANG file.
2) The constructor was modified to accept the raw enum name as a parameter, and store
the parameter value in the name instance variable.
3) An accessor was added called "getName()", which exposes the name instance variable.

A corresponding change in yangtools changes the EnumEffectiveStatementImpl.getName()
definition to expose the "raw" name attribute parsed from the yang file:

https://git.opendaylight.org/gerrit/#/c/39733/

Change-Id: If248039a2b10958ff563ee0799d7b13284570b3c
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>