fixed assertion error in ShardedDOMDataWriteTransactionTest
getIdentifier() method returns String (as a name) and counter in form
of incremented Atomic long in every new instantiation. Before this 
change test exactly expected to have an 0 and then 1 at counter side, 
but if some other test create an instance of 
ShardedDOMDataWriteTransaction before this test, counter will alredy be 
greater than 0 so presumably that is what happend and test is order 
dependant. I removed this dependency and now test is checking for 
non-equality of two differend object identifiers, so does not matter 
what is at counter side at beggining of this test.

Change-Id: I65401ea99bc7d8d94e4ef1fc9b52a1b9909f63fe
Signed-off-by: Peter Nosal <peter.nosal@pantheon.tech>