Update patch set 1
Patch Set 1:

> > How about configuring copyright info via the yang-maven-plugin?
 > By
 > > default we would generate the generic ODL copyright as you did.
 > 
 > Expanding a bit, we could add an optional <copyright> element to
 > the yang-maven-plugin config.

I like that idea, but I'll wait for others to chime in.  GeneratorJavaFile is instantiated by CodeGeneratorImpl, which is MavenProjectAware (so already contains ability to set/get license info).  I think we could probably modify the GeneratorJavaFile to take a maven context and then pass that on to the code generators.  That said, I may be overlooking something.  I will wait to see what Robert and others think about your approach, Tom.  Thanks for the feedback.

Patch-set: 1