Remove project local Xtend dependencyManagement and Maven plugin
Because we now have this available in odl-parent.  Note that the
configuration there is slightly different and uses
target/generated-source instead of src.  Therefore the clean is no
longer needed.  Also add-source does not seem to be required; I've found
that the Xtend Maven plugin and the Xtend IDE M2E support already cover
this correctly by themselves.

Change-Id: Ic878b7c3d63c06595e61c42f0b729bf3d50861cf
Signed-off-by: Michael Vorburger <vorburger@redhat.com>