Update patch set 27
Patch Set 27: Code-Review-1

There are basically 2 public interfaces for the singleton service functionality, CommonClusterSingletonServiceProvider and ClusterSingletonService. A user implements ClusterSingletonService and registers instances to the CommonClusterSingletonServiceProvider implementation provided by core. 

However, in looking at these interfaces, there are no references to any EOS interfaces and no binding vs DOM concepts. The CommonClusterSingletonServiceProvider implementation uses EOS under the hood but this isn't exposed to users. So why then are there binding and DOM implementations of the singleton service provider? It seems this is not needed - all we need is a ClusterSingletonServiceProvider interface and a single implementation that uses the DOM EOS under the hood (the binding EOS is just an adapter that wraps the DOM EOS).

Patch-set: 27
Label: Code-Review=-1