BUG-865: remove String-based getDataChildByName()
This is follow-up patch to reflect Yangtools changes
and needs to be merged after
https://git.opendaylight.org/gerrit/#/c/42898

Note:
based on information from YT devs, relying on string-based
choice node local-name when lookup by QName failed
was useful when dealing with augmented cases.
This is supposed to be fixed now in YT and MDSAL should
rely on lookup by QName only. See patch:
https://git.opendaylight.org/gerrit/#/c/40158/

Change-Id: I1300f3ee01ad678757d564ae76e23ce2c57c76fb
Signed-off-by: Martin Ciglan <mciglan@cisco.com>