Update patch set 5
Patch Set 5: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/mdsal-distribution-check-carbon/76/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/mdsal-distribution-check-carbon/76/ )

https://jenkins.opendaylight.org/releng/job/mdsal-verify-carbon-mvn33-openjdk8/100/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/mdsal-verify-carbon-mvn33-openjdk8/100/ )

https://jenkins.opendaylight.org/releng/job/mdsal-validate-autorelease-carbon/34/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/mdsal-validate-autorelease-carbon/34/ )

Patch-set: 5
Label: Verified=-1