Update patch set 6
Patch Set 6: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/mdsal-validate-autorelease-carbon/42/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/mdsal-validate-autorelease-carbon/42/ )

https://jenkins.opendaylight.org/releng/job/mdsal-verify-carbon-mvn33-openjdk8/116/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/mdsal-verify-carbon-mvn33-openjdk8/116/ )

https://jenkins.opendaylight.org/releng/job/mdsal-distribution-check-carbon/92/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/mdsal-distribution-check-carbon/92/ )

Patch-set: 6
Label: Verified=-1