Update patch set 1
Patch Set 1:

> how about just 'extends SimpleImmutableEntry

sure, that works for me; done - OK?

I'm slightly hesitant if for a class like this now the correct thing to do would be to still keep a customer equals() re-implemented here.. this is a fairly special case: SimpleImmutableEntry equals checks the type, as any good equals should, but not to SimpleImmutableEntry but it does if (!(o instanceof Map.Entry)) .. which would work even if if didn't re-implement it - but it would mean that SimpleImmutableEntry could be equals an InstanceIdentifierDataObjectPair - is that a Feature, or a bug? ;-)  I have a feeling your preference is going to be to remove equals() and hashCode() from this class in this particular case?

 > Also, the name is quite long and does seem to represent something,
 > which should really have a proper name.

Agreed.. any suggestions for what you would call it, better?

Patch-set: 1