Fix groupId to reflect org.opendaylight.mdsal
The merge job is currently broken because of that. It is failing to 
deploy binding-aggregator because it doesn't have the correct groupId.
See https://jenkins.opendaylight.org/releng/job/mdsal-merge-carbon/23/console
for more details.

This issue was introduced with the clean up of the pom file that occured
here https://git.opendaylight.org/gerrit/#/c/45912/

Change-Id: Ia3cae6215547702467d42d604641ec46667bc488
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>