Update patch set 1
Patch Set 1:

> Pure syntactic sugar hinting at a different take on what interface
 > is really provided. I would like to see a well-documented case how
 > this improves things

Sugar yes, but useful.  If interested, please see https://git.opendaylight.org/gerrit/#/c/46206/ for a, first, example of how we're starting to use this.  Please ignore and don't bark at the use of Xtend there, that is completely unrelated to the idea here.  What we're suggesting here and in https://git.opendaylight.org/gerrit/#/c/46479/ is that there is A LOT of code in ODL which creates LogicalDatastoreType / InstanceIdentifier / DataObject triples, and it would be valuable to have a suitable abstraction for this in the base APIs, in order to be able to write code which returns such triples as one single thing.

 > and I do find things like DatastoreUtil.write(path,
 > object), which internally allocates and submits a transaction quite
 > horrible.

Erm... OK, but that's completely unrelated to what we are proposing and discussing here?  Could you be mixing this discussion up with another Gerrit we're looking at in parallel together?  This one has technically nothing to do at all with that one.  (I just happen to have raised them at the same time.)

Patch-set: 1