Update patch set 1
Patch Set 1:

> Sorry for that long comment, maybe we should open the topic on a
 > mail thread?

Yes we should have a discussion re. transaction management utilities at some point (it's not a top prio. for me... I got other bigger fish to fry before; go ahead if you have the cycles for it!), BUT as I stated above, that's the topic for https://git.opendaylight.org/gerrit/#/c/46335/ & follow-ups, but has nothing to do with this proposal and the related https://git.opendaylight.org/gerrit/#/c/46479/.

Patch-set: 1