[DO NOT MERGE YET] Bug 6859: Cleanup package names for mdsal-binding-generator-api module
Most packages still contain yangtools in the package name in mdsal which seems to be
a by-product of cleaning up of the Java Binding code from YANG Tools to md-sal.

The APIs are copied to new package and old APIs extend the new ones.
Added deprecated annotation to the old api.

Change-Id: Ic120a7ec1f85d8f604aa575047d8bd639630517b
Signed-off-by: Rashmi Pujar <rpujar@inocybe.com>