mdsaltrace utility for debugging
TracingBroker logs 'write' operations
and listener registrations to the md-sal. It logs the instance
identifier path, the objects themselves, as well as the stack trace of
the call invoking the registration or write operation. It works by
operating as a "bump on the stack" between the application and actual
DataBroker, intercepting write and registration calls and writing to the
log.

Change-Id: I715b35d9d7b8234326c2be9556295e6c7e3c4b3d
Signed-off-by: Josh <jhershbe@redhat.com>