Bug 6236: Introduce "mdsal.skip.verbose" property, for build speed
This doesn't "fix" Bug 6236, but works around it - and gives us a low
hanging fruit build performance improvement.

If this change gets accepted, I would make the odlparent -Pq profile set
this to true by default.

In parallel, I'm trying to get a root cause fix in a future Xtend
version. (At which point this could be removed again.)

Change-Id: I108f401fa2fbca6f17a8b23cb098910c5a6021fb
Signed-off-by: Michael Vorburger <vorburger@redhat.com>