Clean up apparently dead (and not thread safe) code
YangTemplate's private static Module is necessarily always null, so this
is dead code, and it should be perfectly safe, and clearer, to remove
this.

Full disclosure: I actually have absolutely no idea what writeIdentityNs
was originally good for, but don't think I'm breaking anything with this
clean up, because the way the code is now there is no way that this can
ever be anything else than null.

Change-Id: Ib08112c836efdc1c72443a5f93549a96f954dae8
Signed-off-by: Michael Vorburger <vorburger@redhat.com>