Update patch set 1
Patch Set 1:

> That's because I didn't expect recheck to work already after it had
 > just failed!

That's fine. The validate-autorelease job failure is an intermittent issue. I don't know why sometimes Jenkins does not correctly load the ssh-agent but a 2nd try always works. Unfortunately we're now failing on something else so I guess we'll have to wait for that to get fixed fix first.

Patch-set: 1