AssertDataObjects with LOG of diff
When running failing tests in CI on build (or simply local mvn CLI),
then the ComparisonFailure for large expected/actual objects, such as
flows, is very difficult to make sense of.

In the IDE one doesn't have this problem, because the JUnit view of IDEs
will interpret an ComparisonFailure and show a very clear side-by-side
comparison of expected VS actual.

This change introduces a LOG, in case of AssertDataObjects failure,
which prints a diff for the ComparisonFailure, which will make analysing
AssertDataObjects based test failures much easier on CI build.

Change-Id: I666fbceeb794f7b810a11c2efa8e33a51afc1b2d
Signed-off-by: Michael Vorburger <vorburger@redhat.com>