Bug 7271: Modify AbstractDOMStoreTreeChangePublisher to batch candidates
The AbstractDOMStoreTreeChangePublisher was modified to keep track of
DataTreeCandidates per listener via a Map in order to notify listeners
of all DataTreeCandidate changes in one event. This allows the
initial event on registration to send all list entry nodes. The
previous notifyListeners abstract method was removed in lieu of a
new notifyListener that takes a listener registration and the list
of DataTreeCandidates

Change-Id: I02f14f202a46bbf9afe8de1941701849a41a431c
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>