Update patch set 2
Patch Set 2: Code-Review-1

Please don't. InstanceIdentifier as a concept does not have a value, hence it is wrong to add this method to the public interface.

The correct fix should be delivered to the code generator, so it does not emit getValue().

Patch-set: 2
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1