Bug 7479: Add IETF models from Honeycomb/VBD
The idea is to have just one variant of ietf routing-related
models in ODL, so that Honeycomb/Vbd and Bier projects can share it.

Due to Bug 7535, ietf-routing has changed content.
Most of the content is copied from Honeycomb/Vbd
(hc2vpp-ietf-routing@2014-05-24.yang)
but the module name and the namespace was changed atop of that.
Similarly, ietf-ip and ietf-nat have have revision-date added.

ietf-ipv4-unicast-routing and -ipv6- models are not included due to:
"yangtools bug(unable to proceed augmentation
in separate model of augmentation in this model)"

+ Modules in ietf/pom.xml reorganized.
+ Newly added Yang file namess contain their revision.

Change-Id: I8381464806f1e0473574a632a3ec1bc0c4cfe98f
Signed-off-by: Vratko Polak <vrpolak@cisco.com>