BUG-3128: make context reference mandatory
It does not make quite sense to have an optional context
reference: the RPC is either routed (action, with a reference)
or not. If the leaf is not mandatory, there is is a mismatch
between providers and consumers with regard to what is it
that invoking an action as an RPC really means.

Change-Id: I4f86acdc205192e3771fe9c480a71f38e1fe3e6f
Signed-off-by: Robert Varga <rovarga@cisco.com>