Add new DOMDataTreeCommitCohort#canCoomit method
If a cohort registers for yang list entries, there may be multiple
list entry modifications in a transaction thus the canCommit method
needs to take a Collection of DOMDataTreeCandidates for consistency.

For avoid breaking current implementations, I added the new method
with a default implementation that invokes the single DOMDataTreeCandidate
canCommit method with a warning that the new canCommit method should
be implemented. The single canCommit method is now deprecated with a
default implementation that logs an error and returns success.

Change-Id: If83d380a0ca8ba55be9e99b3d1da4d309e9a6487
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>