Update patch set 3
Patch Set 3:

> I would prefer to apply this change on binding v2 only, which means
 > twirl *.txt templates

Why?  I understand the wish for backward compatibility in general, but in this case that seems over reaching to me - I can't possibly think of anything that hypothetically break just because there is a new annotation on the generated code... or can you imagine a scenario where this could be "risky" which I may be missing?

Consider this more like something like an ... formatting change, or something.  Especially as it's an annotation from java.lang.*  - so it's not even possible that this could theoretically break anything if you have a wrong classpath - it really cannot.  So why the hesitation?

Patch-set: 3
Label: Verified=0