Update patch set 3
Patch Set 3:

> I might be bit more conservative here.

I got that, but you're still not "justifying" why though... ;-)  but you don't have to, of course.  BTW with c/54982 applying a very similar change inspired by this on the binding 1 generator of carbon, the argument to want to be conservative for this change here makes even less sense, to me.

 > Is 8226 supposed to be in both Carbon & master to be consistent?

I don't really care, but perhaps we could take get this at least into master already, for now?  That way if anything unexpectedly is affected or does break (which I really doubt!) we can deal with it there, and not take any risk on carbon - that would be just fine for me.  Perhaps we could then later cherry-pick it to carbon - or never.

Patch-set: 3