Update patch set 18
Patch Set 18:

We are looking to build Carbon RC3 tonight at 23:59 UTC 5/21/2017 assuming there are no blocker bugs.  Can someone cherry pick to stable/carbon and provide an ETA for when it can be merged for stable/carbon branch?

To better assess the impact of this bug and fix, could someone from your team please help us identify the following:

Regression: Is this bug a regression of functionality/performance/feature compared to Boron?

Severity: Could you elaborate on the severity of this bug?  Is this a BLOCKER such that we cannot release Carbon without it?

Is there a workaround such that we can write a release note?

Testing: Could you also elaborate on the testing of this patch?  How extensively has this patch been tested?  Is it covered by any unit tests or system tests?  

Impact: Does this fix impact any dependent projects?

Patch-set: 18