Add ietf-lisp-address-types.yang to models
The ietf-lisp-address-types.yang model was originally part of the
lispflowmapping project. To avoid circular dependencies, it was added to
the honeycomb/vbd project too. That's a less then ideal situation,
leading to some issues which were worked around. The proper solution is
to have a single copy in the OpenDaylight ecosystem, and the right place
for that is the MD-SAL IETF models.

Change-Id: Iaab229b9a792744b1ab0541dbe5a86a24e8904d2
Signed-off-by: Lorand Jakab <lojakab@cisco.com>