Use equals() in AssertDataObjects
Currently, AssertDataObjects.assertEqualBeans() relies entirely on
XtendBeans’ conversions; but what we really want to test is the
equals() implementations — XtendBeans is helpful to produce errors
which are easier to understand and act upon.

This patch tests the beans using equals() first, and only uses
XtendBeans to produce an error message. It also catches cases where
equals() fails but the XtendBeans-generated representations are
equal.

Change-Id: I8412609c54795b3d7c27001f01d3df281203bb78
Signed-off-by: Stephen Kitt <skitt@redhat.com>