Update patch set 4
Patch Set 4: Code-Review-1

(2 comments)

I'm for this.. in fact I've a vague memory that it used to be like this, or perhaps I just wanted to change it to this once, but then either change it back, or could not change it, or yet (but much has changed re this, notably xtendbeans bugs fixed, and equals bugs in genius remedied) - so OK for me in principle.

-1 just re. minor nit comment re. message, AND an important Q:

We'd need to test this against ALL (few..) tests we have in genius and netvirt which use AssertDataObjects (both direct usages and indirect ones, notably via the TestIMdsalApiManager), to make sure that this change does not break them... the +1 from Jenkins on mdsal is not enough.  Have you had a chance to do this?

Patch-set: 4
Label: Code-Review=-1