Fix generating of nested class with the same name as the outer class
Binding v1 generator generates a nested class for a union/bits/enumeration leaf node.
If the parent node of such a leaf has the same name, we end up
generating a java interface that has the same name as the nested class
within it which makes it uncompilable.

Therefore, add a suffix to the name of the nested class in this
particular case, thus avoiding the naming conflict.

JIRA: MDSAL-58
Change-Id: Iba87b41b018bb6932fa74c47c9bfa74993b01d26
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>