Optimized range/length generators
This patch changes the layout of primitive length/range checker code,
so that the formatting string is an invariant.

Since JLS8 section 3.10.5 specifies that all string literals are interned,
this leads to the format string being fully shared. The cost being an
additional argument to String.format() in the error path.

Change-Id: I98f103e9c05a1d68f0f4bc4c6c33146cb654e5bb
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>