Workaround for Findbugs unconfirmed cast
EntityOwnershipChange inherits the getEntity method from
GenericEntityOwnershipChange whose return type is parameterized
to Entity however Findbugs gets confused with

  Entity entity = entityOwnershipChange.getEntity();

and flags a violation for "Unchecked/unconfirmed cast" from
GenericEntity to Entity. I believe this is due to erasure of the
actual return type. To alleviate this, override getEntity() in
EntityOwnershipChange with the actual return type Entity.

Change-Id: I5cf8aa7ab4672c9efb0e13244e698ea9595bfbe5
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>